+++ packages/ui/package.json
@@ -37,12 +37,12 @@
     "typescript": "^4.5.2"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.2.0",
-    "@fortawesome/free-solid-svg-icons": "^6.2.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@fortawesome/fontawesome-svg-core": "6.2.0",
+    "@fortawesome/free-solid-svg-icons": "6.2.0",
+    "@fortawesome/react-fontawesome": "0.2.0",
     "@sovryn/tailwindcss-config": "*",
     "classnames": "^2.3.1",
-    "react-router-dom": "^6.3.0"
+    "react-router-dom": "6.3.0"
   },
   "files": [
     "dist"
