+++ packages/ui/src/1_atoms/Button/Button.types.ts
@@ -1,0 +1,34 @@
+export enum ButtonType {
+  button = 'button',
+  submit = 'submit',
+  reset = 'reset',
+}
+
+export enum ButtonColor {
+  primary = 'primary',
+  secondary = 'secondary',
+  tradeLong = 'tradeLong',
+  tradeShort = 'tradeShort',
+  success = 'success',
+  warning = 'warning',
+  gray = 'gray',
+}
+
+export enum ButtonSize {
+  /** Inline/Minor buttons, height 30px */
+  sm = 'sm',
+  /** Standard buttons, height 40px */
+  md = 'md',
+  /** Minor CTA buttons, height 50px */
+  lg = 'lg',
+  /** Major CTA buttons, height 50px */
+  xl = 'xl',
+}
+
+export enum ButtonStyle {
+  normal = 'normal',
+  inverted = 'inverted',
+  transparent = 'transparent',
+  frosted = 'frosted',
+  link = 'link',
+}
