+++ packages/ui/src/1_atoms/Lead/Lead.tsx
@@ -1,5 +1,5 @@
 import React, { FC, PropsWithChildren } from 'react';
-import styles from './lead.module.css';
+import styles from './Lead.module.css';
 
 export type LeadProps = {
   test?: boolean;
