+++ packages/ui/src/2_molecules/Dropdown/Dropdown.stories.tsx
@@ -9,9 +9,7 @@ export default {
 };
 
 const Template: Story<ComponentProps<typeof Dropdown>> = args => (
-  <>
-    <Dropdown {...args} />
-  </>
+  <Dropdown {...args} />
 );
 
 export const Basic = Template.bind({});
