+++ packages/ui/src/2_molecules/Menu/components/MenuItem/MenuItem.module.css
